+++ src/deploy/function_account_delete.sql
@@ -1,14 +1,14 @@
 BEGIN;
 
 CREATE FUNCTION maevsi.account_delete(
-  "password" TEXT
+  password TEXT
 ) RETURNS VOID AS $$
 DECLARE
   _current_account_id UUID;
 BEGIN
   _current_account_id := current_setting('jwt.claims.account_id')::UUID;
 
-  IF (EXISTS (SELECT 1 FROM maevsi_private.account WHERE account.id = _current_account_id AND account.password_hash = crypt($1, account.password_hash))) THEN
+  IF (EXISTS (SELECT 1 FROM maevsi_private.account WHERE account.id = _current_account_id AND account.password_hash = crypt(account_delete.password, account.password_hash))) THEN
     IF (EXISTS (SELECT 1 FROM maevsi.event WHERE event.created_by = _current_account_id)) THEN
       RAISE 'You still own events!' USING ERRCODE = 'foreign_key_violation';
     ELSE