+++ src/deploy/function_guest_count.sql
@@ -3,7 +3,7 @@ BEGIN;
 CREATE FUNCTION maevsi.guest_count(event_id UUID)
 RETURNS INTEGER AS $$
 BEGIN
-  RETURN (SELECT COUNT(1) FROM maevsi.guest WHERE guest.event_id = $1);
+  RETURN (SELECT COUNT(1) FROM maevsi.guest WHERE guest.event_id = guest_count.event_id);
 END;
 $$ LANGUAGE PLPGSQL STRICT STABLE SECURITY DEFINER;