+++ src/deploy/function_profile_picture_set.sql
@@ -7,11 +7,11 @@ BEGIN
   INSERT INTO maevsi.profile_picture(account_id, upload_id)
   VALUES (
     current_setting('jwt.claims.account_id')::UUID,
-    $1
+    profile_picture_set.upload_id
   )
   ON CONFLICT (account_id)
   DO UPDATE
-  SET upload_id = $1;
+  SET upload_id = profile_picture_set.upload_id;
 END;
 $$ LANGUAGE PLPGSQL STRICT VOLATILE SECURITY INVOKER;