+++ src/deploy/table_account_block_policy.sql
@@ -5,18 +5,19 @@ GRANT SELECT ON TABLE maevsi.account_block TO maevsi_anonymous;
 
 ALTER TABLE maevsi.account_block ENABLE ROW LEVEL SECURITY;
 
--- Only allow account blocking creation with accurate creator.
-CREATE POLICY account_block_insert ON maevsi.account_block FOR INSERT WITH CHECK (
-  maevsi.invoker_account_id() IS NOT NULL
-  AND
+-- Only allow selects by the current account.
+CREATE POLICY account_block_select ON maevsi.account_block FOR SELECT USING (
   created_by = maevsi.invoker_account_id()
 );
 
--- Only show account blockings which are created by the current user or which affect the current user.
-CREATE POLICY account_block_select ON maevsi.account_block FOR SELECT USING (
+-- Only allow inserts by the current account.
+CREATE POLICY account_block_insert ON maevsi.account_block FOR INSERT WITH CHECK (
   created_by = maevsi.invoker_account_id()
-  OR
-  blocked_account_id = maevsi.invoker_account_id()
+);
+
+-- Only allow deletes by the current account.
+CREATE POLICY account_block_delete ON maevsi.account_block FOR DELETE USING (
+  created_by = maevsi.invoker_account_id()
 );
 
 COMMIT;