+++ src/revert/function_account_upload_quota_bytes.sql
@@ -1,5 +1,5 @@
 BEGIN;
 
-DROP FUNCTION maevsi.account_upload_quota_bytes;
+DROP FUNCTION maevsi.account_upload_quota_bytes();
 
 COMMIT;