+++ src/revert/function_event_delete.sql
@@ -1,5 +1,5 @@
 BEGIN;
 
-DROP FUNCTION maevsi.event_delete;
+DROP FUNCTION maevsi.event_delete(UUID, TEXT);
 
 COMMIT;