+++ src/revert/function_event_search.sql
@@ -1,5 +1,5 @@
 BEGIN;
 
-DROP FUNCTION maevsi.event_search;
+DROP FUNCTION maevsi.event_search(TEXT, maevsi.language);
 
 COMMIT;