+++ src/revert/function_trigger_metadata_update.sql
@@ -1,5 +1,5 @@
 BEGIN;
 
-DROP FUNCTION maevsi.trigger_metadata_update;
+DROP FUNCTION maevsi.trigger_metadata_update();
 
 COMMIT;