+++ src/verify/index_guest_contact_id.sql
@@ -1,9 +1,0 @@
-BEGIN;
-
-SELECT 1/COUNT(*)
-FROM   pg_class c
-JOIN   pg_namespace n ON n.oid = c.relnamespace
-WHERE  c.relname = 'idx_guest_contact_id'
-AND    n.nspname = 'maevsi';
-
-ROLLBACK;