+++ src/verify/table_guest.sql
@@ -9,4 +9,8 @@ SELECT id,
        updated_by
 FROM maevsi.guest WHERE FALSE;
 
+SELECT maevsi_test.index_existence(
+  ARRAY ['guest_event_id_contact_id_key', 'idx_guest_updated_by']
+);
+
 ROLLBACK;